Measure it!
Measure it! is an iOS app that allows users to estimate the size of
objects in images.

It was featured in Apple TV Advertisements worldwide and became the #1
Pad app in many stores > https://www.youtube.com/watch?v=DLi9hV11ZXg.

It was last touched in 2009 and will need updating in terms of modern
Obj-C and ARC.